refactor(component): AccountDropdown to persist while user is navigating through different options
